+++ nav2_system_tests/src/behaviors/assisted_teleop/assisted_teleop_behavior_tester.cpp
@@ -246,7 +246,7 @@ void AssistedTeleopBehaviorTester::sendInitialPose()
   pose.pose.pose.orientation.y = 0.0;
   pose.pose.pose.orientation.z = 0.0;
   pose.pose.pose.orientation.w = 1.0;
-  for (int i = 0; i < 35; i++) {
+  for (size_t i = 0; i < 35; i++) {
     pose.pose.covariance[i] = 0.0;
   }
   pose.pose.covariance[0] = 0.08;
